sinefa: Fix the host field truncation, add address and domain fields
The host field was too short to hold long dns names and ipv6
addresses. Increase the size so we don't get truncation.

Also in the json format split the hostname and domain name so that we
don't have to try and parse and decode that host field in userspace
since we will _always_ want to do this.