sinefa: Add mtr clamping to limit latencies (SF-4999)
This adds a new -d/--rtt-clamping option:

| -d, --rtt-clamping         test target first then limit RTT to 75%/150% of target RTT on first/subsequent hops

When set an intial packet will be sent with the TTL set to the
maximum.

If a response is recieved then the RTT will be stored, then the time
limit is set to the rtt * 1.5 (150% of RTT) rounded up the the
nearest second, then the trace proceeeds as normal.

If the RTT is stored then replies for the first hop will be ignored
if they are more than 75% of the stored RTT, and for susequent hops
they will be ignored of they are more than 150% of the stored RTT.

The limits will be rounde dup to 100ms if they end up lower than 100ms.

Note that the timed-out error is now handled, and that errors are now passed
up even if other details could not be decoded. This is becuase the timeout
out error is used when the destination does not respond to trigger the
normal processing.

A new options entry has been added to the json output so you can see
which options were enabled when running the mtr:

  "report": {
    "mtr": {
      "src": "europa2",
      "dst": "uranus.twibble.org",
      "dstAddress": "45.32.40.184",
      "tos": "0x0",
      "psize": "64",
      "bitpattern": "0x00",
      "tests": "2",
      "options": ["rttClamping", "useDns", "dns", "asLookup"]
  ...