sinefa: Add end verification (SF-4999)
Add an option to verify that the end desintation has really been reached:

 | -E, --end-verification     test past destination to make sure it is the end

When enabled it test one TTL past the destination, and if it finds more
hosts it removes the destination and continues on as normal.

This works around a cisco issue that is seen with aussiebroadband.
They are replacing a TTL somewhere with a large value (increment it
to work around a bug, or decrementing twice and going -ve turning in
a larger postive value) which makes the trace end early.  This
detects this and shows a much more realistic trace.