sinefa: Collect DS (differentiated service) details and add to the JSON output (SF-5006)
Collect the DS field, aka the IPv4 TOS field and the IPv6 traffic
class field, from the inner packet in the ICMP replies and add the DS
to the JSON output.

If no DS can be determined then it will not appear in the JSON output.

Example:
  ...
     {
        "count": "10",
        "host": "ce-0-14-0-1.r02.tokyjp05.jp.ce.gin.ntt.net",
        "address": "120.88.54.98",
        "domain": "ce-0-14-0-1.r02.tokyjp05.jp.ce.gin.ntt.net",
        "DS": 0,
        "Loss%": 0,
        "Snt": 1,
        "Last": 131.01,
        "Avg": 131.01,
        "Best": 131.01,
        "Wrst": 131.01,
        "StDev": 0
      },
  ...

The DSCP value is the upper 6 bits of the 8 bit DS field (ie DSCP = 'DS >> 2').