sinefa: Don't consider a timeout as an actual result (SF-5088)
We want the timeout to end the initial RTT period, but don't treat it
like a result that contains an address.  So only pass on errors
without an address if it is a timeout and ignore those time out
errors if it is not related to the intial rtt measurement.