mtr: Add a bind interface option (DEM-16)
Mtr has the following options:

  -I, --interface NAME       use named network interface
  -a, --address ADDRESS      bind the outgoing socket to ADDRESS

There are two issues here:

 1. The interface option is just used to determine which inteface to loop
    up an ip address from, which is then used as the source address.

 2. When using a source address (the interface option just sets the
    source address) while it _binds_ to that address it will only
    work if it was an address that would normally be selected
    (ie, routing says it is a valid source address).

What we want for policy based routing is to force the interface to be
used, so that only IP addresses from that interface are considered.

This change adds a new option:

 -Y, --bind-interface IFNAME Bind the outgoing socket to the specified interface

which forces the outgoing socket to be bound to the named interface,
forcing the use of IP addresses attached to that interface as the
source IP address.