mtr: Add option to send a local port end value (DEM-834)
The -L <port> option can be used to set a local port which is used for TCP
(and UDP). But the same port is used over and over which isn't ideal - if
the socket hasn't closed it may not be able to re-open it to test the next
hop.

So add an additioan -A <port> option that sets an end port.  So the local
port will run through the sequence localport to localportMax and then wrap
back to localport.

Therefore you can still control the local port range, but bound it to a
small fixed range of ports.